Move buildRelativeLabel helper to shared module
